+++ crates/core_arch/src/acle/mod.rs
@@ -119,6 +119,9 @@ mod simd32;
 ))]
 pub use self::simd32::*;
 
+#[cfg(any(target_arch = "aarch64", target_feature = "v7"))]
+pub mod crc;
+
 mod sealed {
     pub trait Dmb {
         unsafe fn __dmb(&self);
