+++ crates/core_arch/src/arm/mod.rs
@@ -25,9 +25,7 @@ mod neon;
 pub use self::neon::*;
 
 #[cfg(any(target_arch = "aarch64", target_feature = "v7"))]
-mod crc;
-#[cfg(any(target_arch = "aarch64", target_feature = "v7"))]
-pub use self::crc::*;
+pub use crate::core_arch::acle::crc::*;
 
 #[cfg(any(target_arch = "aarch64", target_feature = "v7"))]
 mod crypto;
